create custom methods
- Users
A method that returns the 3 most recent posts for a given user.
- Posts
A method that updates the posts counter for a user.
A method which returns the 5 most recent comments for a given post.
- Comments
A method that updates the comments counter for a post.
- Likes
A method that updates the likes counter for a post.
- Test the custom methods by using rails console